feat: Refactor Attendance components; rename Employee Details to EmployeeDetails, update imports, and enhance Navbar routing
