+++ client/src/components/pages/HR/Attendance/EmployeeDetails.jsx
@@ -1,10 +1,11 @@
 import React from "react";
-import Employee from "../../../organisms/Attendance/Employee Details/Employee Details"; // Renamed import
+import Employee from "../../../organisms/Attendance/Employee Details/EmployeeDetails"; // Corrected import path
 import NavBar from "../../../organisms/NavBar/NavBar";
+
 const EmployeeDetails = () => {
   return (
     <>
-      <NavBar Comp={Employee} /> {/* Use the renamed component */}
+      <NavBar Comp={Employee} />
     </>
   );
 };
